#123 Handle Boolean type value, as well treat Number as double by default
And fallback to String instead.

+++ src/main/kotlin/com/akazlou/dynoman/function/Function.kt
@@ -62,8 +62,8 @@ abstract class Function<T> {
             // Boolean type
             arg.toBoolean()
         } else {
-            // Long type
-            arg.toLong()
+            // Double type
+            arg.toDouble()
         }
     }
